+++ HISTORY.rst
@@ -28,6 +28,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * Add EAN13 to PO report lines
+* Ported from 7.0 : use refund description in refund wizard
 
 **Bugfixes**